Update summary api

+++ modules/lmscan-backend/src/main/scala/io/leisuremeta/chain/lmscan/backend/entity/Summary.scala
@@ -6,6 +6,6 @@ final case class Summary(
     blockNumber: Long,
     totalAccounts: Long,
     createdAt: Long,
-    totalTxSize: Long,
-    total_balance: Long,
+    totalTxSize: BigDecimal,
+    total_balance: BigDecimal,
 )
